refactor(automation): move bound settlement policy into runtime
