+++ crates/tracedecay-automation-runtime/src/automation/effect_runtime.rs
@@ -11,6 +11,7 @@ pub mod recovery_index;
 #[cfg(not(any(test, feature = "test-helpers")))]
 mod recovery_index;
 pub mod retirement;
+pub mod settlement;
 pub mod terminal;
 
 pub use authority::finalize_terminal_housekeeping;
